multiagent/pyproject.toml
[project]
name = "qwen-langgraph-demo"
version = "0.1.0"
description = "Local LangGraph demo that uses a remote vLLM server over an SSH tunnel."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=1.0.0",
    "langgraph>=0.6.0",
    "openai>=2.0.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]

[project.scripts]
qwen-langgraph-demo = "qwen_langgraph_demo.main:main"
test-vllm-endpoint = "qwen_langgraph_demo.test_endpoint:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
